+++ R/add_l2.R
@@ -16,10 +16,10 @@
 #' @examples
 #' Leidig2018 %>% add_l2(Leidig2018_l2)
 #' @export
-add_l2 <- function(scdf, 
-                   data_l2, 
+add_l2 <- function(scdf,
+                   data_l2,
                    cvar = "case") {
-  
+
   if (is.null(names(scdf))) {
     stop("scdf must have casenames.")
   }
@@ -27,28 +27,28 @@ add_l2 <- function(scdf,
   if (any(is.na(names(scdf)))) {
     stop("scdf has missing casename(s).")
   }
-  
-  for(i in seq_along(scdf)) {
-    
+
+  for (i in seq_along(scdf)) {
+
     casename <- names(scdf)[i]
     id <- which(data_l2[[cvar]] == casename)
-    
+
     if (length(id) > 1) {
       stop("Multiple matches for a casename in the L1 dataset for ", casename)
     }
-    
+
     if (length(id) == 0) {
       warning("No matching case in L2 data found for ", casename)
     }
-    
+
     if (length(id) == 1) {
       scdf[[i]] <- cbind(
-        scdf[[i]], 
-        data_l2[id, -which(names(data_l2) == cvar), drop = FALSE], 
+        scdf[[i]],
+        data_l2[id, -which(names(data_l2) == cvar), drop = FALSE],
         row.names = NULL
       )
     }
   }
-  
+
   scdf
 }