+++ R/batch_apply.R
@@ -3,8 +3,8 @@
 #' This function applies a given function to each case of a multiple case scdf,
 #' returning a list of the output of each function call.
 #'
-#' @param scdf    A list of inputs to apply the function to.
-#' @param fn      The function to apply to each element. Use a `.` as a
+#' @param scdf A list of inputs to apply the function to.
+#' @param fn  The function to apply to each element. Use a `.` as a
 #'   placeholder for the scdf (e.g. `describe(.)`).
 #' @param simplify If simplify is TRUE and `fn` returns a vector of values,
 #'   `batch_apply` will return a data frame case names.
@@ -17,18 +17,19 @@
 batch_apply <- function(scdf, fn, simplify = FALSE) {
   fn <- substitute(fn)
   out <- vector("list", length(scdf))
-  for(i in seq_along(scdf)) {
-    . <- scdf[i]
-    out[[i]] <- eval(fn)
+  for (i in seq_along(scdf)) {
+    data <- list(. = scdf[i])
+    out[[i]] <- eval(fn, envir = data)
   }
   names(out) <- names(scdf)
-  
+
   if (simplify) {
     out <- as.data.frame(do.call(rbind, out))
-    out$case <- names(scdf)
+    out$case <- rep(names(scdf), each = nrow(out) / length(names(scdf)))
+    out$rownames <- rownames(out)
     rownames(out) <- NULL
-    out <- out[, c(ncol(out), 1:(ncol(out) - 1))]
+    out <- out[, c(ncol(out) - 1, ncol(out), 1:(ncol(out) - 2))]
   }
-  
+
   out
 }