+++ R/nap.R
@@ -3,10 +3,13 @@
 #' The [nap()] function calculates the nonoverlap of all pairs (NAP; Parker &
 #' Vannest, 2009).  NAP summarizes the overlap between all pairs of phase A and
 #' phase B data points.  If an increase of phase B scores is expected, a
-#' non-overlapping pair has a higher phase B data point.  The NAP equals
-#' *number of pairs showing no overlap / number of pairs*.  Because NAP can
-#' only take values between 50 and 100 percent, a rescaled and therefore more
-#' intuitive NAP (0-100\%) is also displayed.
+#' non-overlapping pair has a higher phase B data point. The NAP equals
+#' *number of pairs showing no overlap / number of pairs* where ties are
+#' counted as half non-overlaps.  Because NAP can take values between 0 and 100
+#' percent where values below 50 percent indicate an inverse effect, an nap
+#' rescaled from -100 to 100 percent where negative
+#' values indicate an inverse effect is also displayed (\eqn{nap_{rescaled} = 2
+#' * nap - 100}).
 #'
 #' @inheritParams .inheritParams
 #' @return \item{nap}{A data frame with NAP and additional values for each
@@ -46,36 +49,47 @@ nap <- function(data, dvar, pvar,
     pairs <- length(values$A) * length(values$B)
     
     if (!decreasing) {
-      pos <- pairs -
-        sum(unlist(lapply(values$A, function(x) x >= values$B)))
+      pos <- sum(unlist(lapply(values$A, function(x) values$B > x)))
     }
+    
     if (decreasing) {
-      pos <- pairs -
-        sum(unlist(lapply(values$A, function(x) x <= values$B)))
+      pos <- sum(unlist(lapply(values$A, function(x) values$B < x)))
     }
     
     ties <- sum(unlist(lapply(values$A, function(x) x == values$B)))
-    nap <- (pos + (0.5 * ties)) / pairs
     
-    test <- wilcox.test(values$A, values$B,
-                        alternative = if (decreasing) "greater" else "less",
-                        exact = FALSE
+    non_overlaps <- pos + (0.5 * ties)
+    
+    nap <- non_overlaps / pairs
+    
+    test <- wilcox.test(
+      values$A, values$B,
+      alternative = if (decreasing) "greater" else "less",
+      exact = FALSE
     )
-    
-    list(
+    #nap <- (pairs - test$statistic) / pairs
+    d <- 3.464 * (1 - sqrt((1 - nap) / 0.5))
+    r <- d / sqrt(d^2 + 4)
+    data.frame(
       NAP = nap * 100,
-      Rescaled = 2 * (nap * 100) - 100,
+      "NAP Rescaled" = 2 * (nap * 100) - 100,
       Pairs = pairs,
+      "Non-overlaps" = non_overlaps,
       Positives = pos,
       Ties = ties,
       w = test$statistic,
-      p = test$p.value
+      p = test$p.value,
+      d = d,
+      "R²" = r^2,
+      check.names = FALSE
     )
-    
+ 
   }  
   
   x <- lapply(data, .nap)
   nap <- do.call(rbind, x)
+  nap <- cbind(Case = casenames, nap)
+  rownames(nap) <- NULL
   
   out <- list(nap = nap)
   class(out) <- c("sc_nap")