+++ inst/shiny/server.R
@@ -9,17 +9,19 @@ server <- function(input, output, session) {
 
   ## Render ----
   my_scdf <- reactiveVal()
+  
   scdf_render <- reactive({
-    
-    if (!inherits(my_scdf(), "scdf")) validate(res$msg$no_case)
     
     if (input$scdf_output_format == "Summary"){
       output$scdf_output <- renderPrint({
+        if (!inherits(my_scdf(), "scdf")) validate(res$msg$no_case)
+        if (identical(length(my_scdf()), 0)) validate(res$msg$no_case)
         do.call("summary", list(my_scdf()))
       })
     } else if (input$scdf_output_format == "Syntax") {
       output$scdf_output <- renderPrint({
-        req(inherits(my_scdf(), "scdf"))
+        if (!inherits(my_scdf(), "scdf")) validate(res$msg$no_case)
+        if (identical(length(my_scdf()), 0)) validate(res$msg$no_case)
         do.call("convert", list(
           my_scdf(), inline = as.logical(input$scdf_syntax_phase_structure)
         ))
@@ -35,9 +37,13 @@ server <- function(input, output, session) {
     if (input$scdf_example != "(none)") {
       my_scdf(paste0("scan::", input$scdf_example) |> str2lang() |> eval())
       scdf_render()
-      output$scdf_messages <- renderPrint(
+      output$load_messages <- renderPrint(
         cat(paste0("loaded example ", input$scdf_example))
       )
+      output$load_output <- renderPrint({
+        do.call("summary", list(my_scdf()))
+      })
+      output$scdf_messages <- renderPrint(cat(""))
     } else {
       my_scdf(NULL)
     }
@@ -60,12 +66,16 @@ server <- function(input, output, session) {
     }
 
     if (!inherits(new, "scdf")) {
-      output$scdf_messages <- renderText(
+      output$load_messages <- renderText(
         "Sorry,\n the file you tried to load is not a valid scdf file.")
     } else {
       my_scdf(new)
       scdf_render()
-      output$scdf_messages <- renderPrint(cat(paste0("loaded file successfully")))
+      output$load_messages <- renderPrint(cat(paste0("loaded file successfully")))
+      output$load_output <- renderPrint({
+        do.call("summary", list(my_scdf()))
+      })
+      output$scdf_messages <- renderPrint(cat(""))
     }
 
   })
@@ -185,8 +195,8 @@ server <- function(input, output, session) {
     scdf_render()
   })
 
-  ## remove all cases --------
-  observeEvent(input$remove_all, {
+  ## clear cases --------
+  observeEvent(input$clear_cases, {
     my_scdf(NULL)
     output$scdf_messages <- renderPrint(cat("Cleared cases"))
     scdf_render()
@@ -246,8 +256,8 @@ server <- function(input, output, session) {
     if (length(syntax)>1) {
       syntax <- syntax[-1]
       syntax <- paste0(
-        "scdf %>%\n  ",
-        paste0(syntax, collapse = " %>%\n  ")
+        "scdf", res$pipe_br, " ",
+        paste0(syntax, collapse = paste0(res$pipe_br, " "))
       )
 
     }
@@ -310,10 +320,11 @@ server <- function(input, output, session) {
   output$stats_html <- renderUI({
     results <- calculate_stats()
     print_args <- input$stats_print_arguments
+    flip <- paste0("flip = ", input$stats_export_flip)
     if (print_args != "") {
       print_args <- paste0(", ", print_args)
-      call<- paste0("export(results, ", print_args, ")")
-    } else call <- "export(results)"
+      call<- paste0("export(results, ", flip, ", ", print_args, ")")
+    } else call <- paste0("export(results,", flip ,")")
     tryCatch(
       str2lang(call) |> eval() |> HTML(),
       error = function(e)
@@ -348,9 +359,12 @@ server <- function(input, output, session) {
       "data", "scdf", "data.l2", "offset", "lag.max",
       "graph", "output", "...")
     )
-    args <- args[id]
-    values <- values[id]
-    list(names = args, values = values)
+    if (input$func == "mplm") {
+      id <- c(which(args == "dvar") , id)
+      values[which(args == "dvar")] <- scan:::dv(transformed())
+    }
+    
+    list(names = args[id], values = values[id])
   })
 
   output$stats_arguments <- renderUI({
@@ -472,7 +486,6 @@ server <- function(input, output, session) {
         }
         out <- str2lang(call) |> eval()
         kableExtra::save_kable(out, file)
-        #writeLines(out, con = file)
       }
       
     }
@@ -488,7 +501,7 @@ server <- function(input, output, session) {
       plot_args <- trimws(input$plot_arguments)
       plot_args <- gsub("\n+", "\n", plot_args)
       call <- paste0(
-        call, "%>% ", gsub("\n", " %>% ", plot_args)
+        call, res$pipe, gsub("\n", res$pipe, plot_args)
       )
     }
     call <- paste0("print(",call,")")
@@ -548,7 +561,7 @@ server <- function(input, output, session) {
     call <- paste0("scplot(scdf)")
     if (trimws(input$plot_arguments) != "") {
       call <- paste0(
-        call, "%>%\n  ", gsub("\n", " %>%\n  ", trimws(input$plot_arguments))
+        call, res$pipe_br, " ", gsub("\n", paste0(res$pipe_br, " "), trimws(input$plot_arguments))
       )
     }
     output$plot_syntax <- renderPrint({
@@ -601,13 +614,13 @@ server <- function(input, output, session) {
     )  
     
     ci <- input$pt_ci
-
+   
     syntax <- paste0(
-      syntax, " %>% \n",
+      syntax, res$pipe,
       "  power_test(method = ", deparse(input$pt_method), ", ",
       "effect = ", deparse(input$pt_effect), ", ",
-      "n_sim = ", input$pt_n, ", ",
-      "ci = ", ci,
+      "n_sim = ", input$pt_n, 
+      if (!identical(ci, "")) paste0(", ci = ", ci),
       ")"
     )
     
@@ -632,8 +645,8 @@ server <- function(input, output, session) {
       #"s = ", input$design_s, ", ",
       "rtt = ", input$design_rtt, ", ",
       "distribution = ", deparse(input$design_distribution),
-      "\n) %>% ",
-      "random_scdf() %>% ",
+      "\n)", res$pipe,
+      "random_scdf()", res$pipe,
       "scplot()"
     )
     
@@ -677,7 +690,12 @@ server <- function(input, output, session) {
   # quit app -----
   observeEvent(input$navpage, {
     if (input$navpage == "Quit") {
+      session$sendCustomMessage(
+        type = "closeWindow", 
+        list(message = "window.close();")
+      )
       stopApp()
     }
   })
+  
 }