+++ NEWS.md
@@ -1,3 +1,12 @@
+# scan 0.60.9999
+
+- Corrected name of level-effect predictor for regressions when the phase variable is not named "phase"
+
+## Changes
+
+- `nap()`: added Cohen's d and R-Squared effects.
+- `export()`: select argument for `nap`; new for `pem()`, `pet()`
+
 # scan 0.60.0
 
 ## New function
@@ -16,7 +25,7 @@
 - Added power-analyses
 - Added settings
 - Extended save options
-- Various changes and optimizations of the ui
+- Various changes and optimization of the ui
 
 # scan 0.59.0